Chore: Move h3 tests into seperate test file
